Added equals method to Patterns and moved the test for patterns
