+++ src/test/java/operations/patterns/SingleOccurrenceValuePatternTest.java
@@ -1,4 +1,4 @@
-package operations.coding;
+package operations.patterns;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
